+++ src/settlers/move/game-logic/roll.service.ts
@@ -1,7 +1,8 @@
-import { BadRequestException, ForbiddenException, Injectable } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { BuildingService } from '../../building/building.service';
 import { Tile } from '../../map/map.schema';
 import { MapService } from '../../map/map.service';
+import { Player } from '../../player/player.schema';
 import { PlayerService } from '../../player/player.service';
 import { ResourceType, TILE_RESOURCES } from '../../shared/constants';
 import { cubeCorners } from '../../shared/hexagon';
@@ -42,7 +43,9 @@ export class RollService {
 
   async roll(gameId: string, userId: string, move: CreateMoveDto): Promise<Move> {
     const roll = this.d6() + this.d6();
-    if (roll !== 7) {
+    if (roll === 7) {
+      await this.roll7(gameId);
+    } else {
       await this.rollResources(gameId, roll);
     }
 
@@ -55,38 +58,88 @@ export class RollService {
     });
   }
 
+  private async roll7(gameId: string): Promise<void> {
+    const players = await this.playerService.findAll(gameId, {active: false});
+    await Promise.all(players.map(p => this.stealResources(p)));
+  }
+
+  private async stealResources(player: Player) {
+    const resources = { ...player.resources };
+    let total = Object.values(resources).sum();
+    if (total <= 7) {
+      return;
+    }
+
+    const keys = Object.keys(resources) as ResourceType[];
+    const stealCount = Math.floor(total / 2);
+
+    for (let i = 0; i < stealCount; i++) {
+      let rand = Math.randInt(total);
+      for (const key of keys) {
+        const amount = resources[key];
+        if (!amount) {
+          continue;
+        }
+
+        if (rand >= amount) {
+          rand -= amount;
+          continue;
+        }
+
+        resources[key] = amount - 1;
+        total--;
+        break;
+      }
+    }
+
+    await this.playerService.update(player.gameId, player.userId, {
+      resources,
+    });
+  }
+
+
   async rob(gameId: string, userId: string, move: CreateMoveDto): Promise<Move> {
     if (!move.rob) {
       throw new BadRequestException('Missing rob property');
     }
 
     const { target: targetId, ...robber } = move.rob;
-    const target = await this.playerService.findOne(gameId, targetId);
-    if (!target) {
-      throw new BadRequestException('Target player does not exist');
+
+    if (targetId) {
+      const target = await this.playerService.findOne(gameId, targetId);
+      if (!target) {
+        throw new NotFoundException(targetId);
+      }
+
+      const buildings = await this.buildingService.findAll(gameId, {
+        owner: targetId,
+        $or: cubeCorners(move.rob),
+      });
+      if (!buildings.length) {
+        throw new ForbiddenException('The target player has no buildings adjacent to the tile');
+      }
+
+      const resources = Object.keys(target.resources).filter(k => target.resources[k as ResourceType]! > 0);
+      if (!resources.length) {
+        throw new BadRequestException('The target player has no resources');
+      }
+
+      const randomResource = resources[Math.randInt(resources.length)];
+      await Promise.all([
+        this.updateResources(gameId, targetId, { [randomResource]: -1 }),
+        this.updateResources(gameId, userId, { [randomResource]: +1 }),
+      ]);
+    } else {
+      const buildings = await this.buildingService.findAll(gameId, {
+        owner: {$ne: userId},
+        $or: cubeCorners(move.rob),
+      });
+      if (buildings.length) {
+        throw new ForbiddenException('There are buildings adjacent to the tile - you must specify a target player');
+      }
     }
 
-    const buildings = await this.buildingService.findAll({
-      gameId,
-      owner: targetId,
-      $or: cubeCorners(move.rob),
-    });
-    if (!buildings.length) {
-      throw new ForbiddenException('The target player has no buildings adjacent to the tile');
-    }
-
-    const resources = Object.keys(target.resources).filter(k => target.resources[k as ResourceType]! > 0);
-    if (!resources.length) {
-      throw new BadRequestException('The target player has no resources');
-    }
-
-    const randomResource = resources[Math.randInt(resources.length)];
-    await Promise.all([
-      this.updateResources(gameId, targetId, { [randomResource]: -1 }),
-      this.updateResources(gameId, userId, { [randomResource]: +1 }),
-      this.stateService.update(gameId, { robber }),
-    ]);
-
+    await this.stateService.update(gameId, { robber });
     return this.moveService.create({
       ...move,
       building: undefined,
@@ -119,8 +172,7 @@ export class RollService {
       return;
     }
 
-    const adjacentBuildings = await this.buildingService.findAll({
-      gameId,
+    const adjacentBuildings = await this.buildingService.findAll(gameId, {
       $or: cubeCorners(tile),
     });
     for (const building of adjacentBuildings) {